Add a name parameter to signups
This adds a name parameter users can provide when signing up. The
goal here is to be able to go from affiliate address -> name, not
the other way around. We don't check uniqueness of names; multiple
addresses could have the same name. The main intent of this is so
that tools that report which dex a trade came from can report the
affiliate based on the provided name.